Move from script to service
